server/app/services/team_service.py
from django.shortcuts import get_object_or_404
from .models import Attendance, LeaveBalance

class TeamService:
    @staticmethod
    def get_team_attendance(manager_id):
        # Logic to retrieve team attendance based on manager_id
        # This will include querying Attendance model and filtering based on employee's manager
        pass

    @staticmethod
    def get_team_leave_requests(manager_id):
        # Logic to retrieve leave requests based on manager_id
        # This will include querying Leave model and filtering based on employee's manager
        pass

    @staticmethod
    def approve_leave_request(manager_id, request_id):
        # Logic to approve leave requests
        # Retrieve the request using request_id and change its status
        pass

    @staticmethod
    def deny_leave_request(manager_id, request_id):
        # Logic to deny leave requests
        # Retrieve the request using request_id and change its status
        pass